automation-scripts/requirements.txt
openai
pyautogui
pillow
